refactor: checking light curate in hook
