+++ src/pages/items-router.tsx
@@ -1,14 +1,12 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { TCRViewProvider } from 'contexts/tcr-view-context'
 import { LightTCRViewProvider } from 'contexts/light-tcr-view-context'
 import loadable from '@loadable/component'
-import _gtcr from '../assets/abis/GeneralizedTCR.json'
-import { useWeb3Context } from 'web3-react'
-import { ethers } from 'ethers'
 import { useParams } from 'react-router'
 import Loading from 'components/loading'
 import useTcrNetwork from 'hooks/use-tcr-network'
 import { NETWORK_STATUS } from 'config/networks'
+import useCheckLightCurate from 'hooks/use-check-light-curate'
 
 const LightItems = loadable(
   () => import(/* webpackPrefetch: true */ './light-items/index'),
@@ -25,30 +23,9 @@ const Items = loadable(
 )
 
 const ItemsRouter = () => {
-  const { tcrAddress } = useParams()
-  const [isLightCurate, setIsLightCurate] = useState()
-  const { library, active } = useWeb3Context()
+  const { tcrAddress } = useParams<{ tcrAddress: string }>()
+  const isLightCurate = useCheckLightCurate()
   const { networkStatus } = useTcrNetwork()
-
-  useEffect(() => {
-    ;(async () => {
-      try {
-        if (!active) return
-        const tcr = new ethers.Contract(tcrAddress, _gtcr, library)
-
-        // Call a function only available on GTCR Classic. If
-        // it throws, its not a light curate instance.
-        await tcr.itemCount()
-        setIsLightCurate(false)
-        // eslint-disable-next-line no-unused-vars
-      } catch (err) {
-        console.info(
-          `Contract call used to verify if this is a Light Curate instance. Ignore exception.`
-        )
-        setIsLightCurate(true)
-      }
-    })()
-  }, [active, library, tcrAddress])
 
   if (
     typeof isLightCurate === 'undefined' ||
